Topics Fragment pulling in all modules

+++ app/src/main/java/uk/airbyte/fwc/model/Module.java
@@ -61,7 +61,7 @@ public class Module extends RealmObject{
     @Expose
     private int currentWindow;
     @Expose
-    private Long playerPosition;
+    private int playerPosition;
 
     /**
      * No args constructor for use in serialization
@@ -73,7 +73,7 @@ public class Module extends RealmObject{
     public Module(String id, String type, String displayOrder, String name, String introduction,
                   String description, String notes, Presenter presenter, Media media, RealmList<String> bullets,
                   Category category, Topic topic, Boolean free, Boolean canView, Boolean favourited, long lastViewed,
-                  int currentWindow, Long playerPosition) {
+                  int currentWindow, int playerPosition) {
         super();
         this.id = id;
         this.type = type;
@@ -231,11 +231,11 @@ public class Module extends RealmObject{
         this.currentWindow = currentWindow;
     }
 
-    public Long getPlayerPosition() {
+    public int getPlayerPosition() {
         return playerPosition;
     }
 
-    public void setPlayerPosition(Long playerPosition) {
+    public void setPlayerPosition(int playerPosition) {
         this.playerPosition = playerPosition;
     }}
